move to inference dir
