+++ tests/V2/Parsing/FailedInferenceResponseTest.php
@@ -5,17 +5,8 @@ declare(strict_types=1);
 namespace V2\Parsing;
 
 use DateTimeImmutable;
-use Mindee\Geometry\Point;
-use Mindee\Input\LocalResponse;
-use Mindee\V2\Parsing\Error\ErrorItem;
 use Mindee\V2\Parsing\Error\ErrorResponse;
-use Mindee\V2\Parsing\FailedInferenceResponse;
-use Mindee\V2\Parsing\Inference\Field\FieldConfidence;
-use Mindee\V2\Parsing\Inference\Field\ListField;
-use Mindee\V2\Parsing\Inference\Field\ObjectField;
-use Mindee\V2\Parsing\Inference\Field\SimpleField;
-use Mindee\V2\Parsing\Job\JobResponse;
-use Mindee\V2\Product\Extraction\ExtractionResponse;
+use Mindee\V2\Parsing\Inference\FailedInferenceResponse;
 use PHPUnit\Framework\TestCase;
 use TestingUtilities;
